refactor: process sync actions out of socket
